+++ aern2-mp/src/AERN2/MP/WithCurrentPrec/Comparisons.hs
@@ -34,18 +34,18 @@ import AERN2.MP.WithCurrentPrec.Type
 instance
     (HasEqAsymmetric t1 t2, p1 ~ p2)
     =>
-    HasEqAsymmetric (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) 
+    HasEqAsymmetric (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) 
     where
-    type EqCompareType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = EqCompareType t1 t2
+    type EqCompareType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = EqCompareType t1 t2
     equalTo = lift2P equalTo
     notEqualTo = lift2P notEqualTo
 
 instance
     (HasOrderAsymmetric t1 t2, p1 ~ p2)
     =>
-    HasOrderAsymmetric (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) 
+    HasOrderAsymmetric (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) 
     where
-    type OrderCompareType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = OrderCompareType t1 t2
+    type OrderCompareType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = OrderCompareType t1 t2
     greaterThan = lift2P greaterThan
     lessThan = lift2P lessThan
     geq = lift2P geq
@@ -54,17 +54,17 @@ instance
 instance
     (CanAbs t)
     =>
-    CanAbs (WithCurrentPrec t p)
+    CanAbs (WithCurrentPrec p t)
     where
-    type AbsType (WithCurrentPrec t p) = WithCurrentPrec (AbsType t) p
+    type AbsType (WithCurrentPrec p t) = WithCurrentPrec p (AbsType t)
     abs = lift1 abs
 
 instance
     (CanMinMaxAsymmetric t1 t2, p1 ~ p2)
     =>
-    CanMinMaxAsymmetric (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) 
+    CanMinMaxAsymmetric (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) 
     where
-    type MinMaxType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = WithCurrentPrec (MinMaxType t1 t2) p1
+    type MinMaxType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = WithCurrentPrec p1 (MinMaxType t1 t2)
     min = lift2 min
     max = lift2 max
 
@@ -79,36 +79,36 @@ $(declForTypes
     instance
       (CanMinMaxAsymmetric a $t)
       => 
-      CanMinMaxAsymmetric (WithCurrentPrec a p) $t
+      CanMinMaxAsymmetric (WithCurrentPrec p a) $t
       where
-      type MinMaxType (WithCurrentPrec a p) $t = WithCurrentPrec (MinMaxType a $t) p
+      type MinMaxType (WithCurrentPrec p a) $t = WithCurrentPrec p (MinMaxType a $t)
       min = lift1T min
       max = lift1T max
 
     instance
       (CanMinMaxAsymmetric a (CN $t))
       => 
-      CanMinMaxAsymmetric (WithCurrentPrec a p) (CN $t)
+      CanMinMaxAsymmetric (WithCurrentPrec p a) (CN $t)
       where
-      type MinMaxType (WithCurrentPrec a p) (CN $t) = WithCurrentPrec (MinMaxType a (CN $t)) p
+      type MinMaxType (WithCurrentPrec p a) (CN $t) = WithCurrentPrec p (MinMaxType a (CN $t))
       min = lift1T min
       max = lift1T max
 
     instance
       (CanMinMaxAsymmetric $t a)
       => 
-      CanMinMaxAsymmetric $t (WithCurrentPrec a p)
+      CanMinMaxAsymmetric $t (WithCurrentPrec p a)
       where
-      type MinMaxType $t (WithCurrentPrec a p) = WithCurrentPrec (MinMaxType $t a) p
+      type MinMaxType $t (WithCurrentPrec p a) = WithCurrentPrec p (MinMaxType $t a)
       min = liftT1 min
       max = liftT1 max
 
     instance
       (CanMinMaxAsymmetric (CN $t) a)
       => 
-      CanMinMaxAsymmetric (CN $t) (WithCurrentPrec a p)
+      CanMinMaxAsymmetric (CN $t) (WithCurrentPrec p a)
       where
-      type MinMaxType (CN $t) (WithCurrentPrec a p) = WithCurrentPrec (MinMaxType (CN $t) a) p
+      type MinMaxType (CN $t) (WithCurrentPrec p a) = WithCurrentPrec p (MinMaxType (CN $t) a)
       min = liftT1 min
       max = liftT1 max
 
@@ -117,36 +117,36 @@ $(declForTypes
     instance
       (HasEqAsymmetric a $t)
       => 
-      HasEqAsymmetric (WithCurrentPrec a p) $t
+      HasEqAsymmetric (WithCurrentPrec p a) $t
       where
-      type EqCompareType (WithCurrentPrec a p) $t = EqCompareType a $t
+      type EqCompareType (WithCurrentPrec p a) $t = EqCompareType a $t
       equalTo = lift1TP (==)
       notEqualTo = lift1TP (/=)
 
     instance
       (HasEqAsymmetric a (CN $t))
       => 
-      HasEqAsymmetric (WithCurrentPrec a p) (CN $t)
+      HasEqAsymmetric (WithCurrentPrec p a) (CN $t)
       where
-      type EqCompareType (WithCurrentPrec a p) (CN $t) = EqCompareType a (CN $t)
+      type EqCompareType (WithCurrentPrec p a) (CN $t) = EqCompareType a (CN $t)
       equalTo = lift1TP (==)
       notEqualTo = lift1TP (/=)
 
     instance
       (HasEqAsymmetric $t a)
       =>
-      HasEqAsymmetric $t (WithCurrentPrec a p)
+      HasEqAsymmetric $t (WithCurrentPrec p a)
       where
-      type EqCompareType $t (WithCurrentPrec a p) = EqCompareType $t a
+      type EqCompareType $t (WithCurrentPrec p a) = EqCompareType $t a
       equalTo = liftT1P (==)
       notEqualTo = liftT1P (/=)
 
     instance
       (HasEqAsymmetric (CN $t) a)
       =>
-      HasEqAsymmetric (CN $t) (WithCurrentPrec a p)
+      HasEqAsymmetric (CN $t) (WithCurrentPrec p a)
       where
-      type EqCompareType (CN $t) (WithCurrentPrec a p) = EqCompareType (CN $t) a
+      type EqCompareType (CN $t) (WithCurrentPrec p a) = EqCompareType (CN $t) a
       equalTo = liftT1P (==)
       notEqualTo = liftT1P (/=)
 
@@ -155,9 +155,9 @@ $(declForTypes
     instance
       (HasOrderAsymmetric a $t)
       => 
-      HasOrderAsymmetric (WithCurrentPrec a p) $t
+      HasOrderAsymmetric (WithCurrentPrec p a) $t
       where
-      type OrderCompareType (WithCurrentPrec a p) $t = OrderCompareType a $t
+      type OrderCompareType (WithCurrentPrec p a) $t = OrderCompareType a $t
       lessThan = lift1TP lessThan
       greaterThan = lift1TP greaterThan
       leq = lift1TP leq
@@ -166,9 +166,9 @@ $(declForTypes
     instance
       (HasOrderAsymmetric a (CN $t))
       => 
-      HasOrderAsymmetric (WithCurrentPrec a p) (CN $t)
+      HasOrderAsymmetric (WithCurrentPrec p a) (CN $t)
       where
-      type OrderCompareType (WithCurrentPrec a p) (CN $t) = OrderCompareType a (CN $t)
+      type OrderCompareType (WithCurrentPrec p a) (CN $t) = OrderCompareType a (CN $t)
       lessThan = lift1TP lessThan
       greaterThan = lift1TP greaterThan
       leq = lift1TP leq
@@ -177,9 +177,9 @@ $(declForTypes
     instance
       (HasOrderAsymmetric $t a)
       =>
-      HasOrderAsymmetric $t (WithCurrentPrec a p)
+      HasOrderAsymmetric $t (WithCurrentPrec p a)
       where
-      type OrderCompareType $t (WithCurrentPrec a p) = OrderCompareType $t a
+      type OrderCompareType $t (WithCurrentPrec p a) = OrderCompareType $t a
       lessThan = liftT1P lessThan
       greaterThan = liftT1P greaterThan
       leq = liftT1P leq
@@ -188,9 +188,9 @@ $(declForTypes
     instance
       (HasOrderAsymmetric (CN $t) a)
       =>
-      HasOrderAsymmetric (CN $t) (WithCurrentPrec a p)
+      HasOrderAsymmetric (CN $t) (WithCurrentPrec p a)
       where
-      type OrderCompareType (CN $t) (WithCurrentPrec a p) = OrderCompareType (CN $t) a
+      type OrderCompareType (CN $t) (WithCurrentPrec p a) = OrderCompareType (CN $t) a
       lessThan = liftT1P lessThan
       greaterThan = liftT1P greaterThan
       leq = liftT1P leq