+++ aern2-mp/src/AERN2/MP/WithCurrentPrec/Field.hs
@@ -34,29 +34,29 @@ import AERN2.MP.WithCurrentPrec.Type
 instance
     (CanAddAsymmetric t1 t2, p1~p2)
     =>
-    (CanAddAsymmetric (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2)) where
-    type AddType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = WithCurrentPrec (AddType t1 t2) p1
+    (CanAddAsymmetric (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2)) where
+    type AddType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = WithCurrentPrec p1 (AddType t1 t2)
     add = lift2 add
 
 instance
     (CanSub t1 t2, p1~p2)
     =>
-    (CanSub (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2)) where
-    type SubType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = WithCurrentPrec (SubType t1 t2) p1
+    (CanSub (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2)) where
+    type SubType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = WithCurrentPrec p1 (SubType t1 t2)
     sub = lift2 sub
 
 instance
     (CanMulAsymmetric t1 t2, p1~p2)
     =>
-    (CanMulAsymmetric (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2)) where
-    type MulType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = WithCurrentPrec (MulType t1 t2) p1
+    (CanMulAsymmetric (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2)) where
+    type MulType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = WithCurrentPrec p1 (MulType t1 t2)
     mul = lift2 mul
 
 instance
     (CanDiv t1 t2, p1~p2)
     =>
-    (CanDiv (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2)) where
-    type DivType (WithCurrentPrec t1 p1) (WithCurrentPrec t2 p2) = WithCurrentPrec (DivType t1 t2) p1
+    (CanDiv (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2)) where
+    type DivType (WithCurrentPrec p1 t1) (WithCurrentPrec p2 t2) = WithCurrentPrec p1 (DivType t1 t2)
     divide = lift2 divide
 
 
@@ -73,129 +73,129 @@ $(declForTypes
     instance
       (CanAddAsymmetric a $t)
       => 
-      CanAddAsymmetric (WithCurrentPrec a p) $t
+      CanAddAsymmetric (WithCurrentPrec p a) $t
       where
-      type AddType (WithCurrentPrec a p) $t = WithCurrentPrec (AddType a $t) p
+      type AddType (WithCurrentPrec p a) $t = WithCurrentPrec p (AddType a $t)
       add = lift1T add
 
     instance
       (CanAddAsymmetric a (CN $t))
       => 
-      CanAddAsymmetric (WithCurrentPrec a p) (CN $t)
+      CanAddAsymmetric (WithCurrentPrec p a) (CN $t)
       where
-      type AddType (WithCurrentPrec a p) (CN $t) = WithCurrentPrec (AddType a (CN $t)) p
+      type AddType (WithCurrentPrec p a) (CN $t) = WithCurrentPrec p (AddType a (CN $t))
       add = lift1T add
 
     instance
       (CanAddAsymmetric $t a)
       => 
-      CanAddAsymmetric $t (WithCurrentPrec a p)
+      CanAddAsymmetric $t (WithCurrentPrec p a)
       where
-      type AddType $t (WithCurrentPrec a p) = WithCurrentPrec (AddType $t a) p
+      type AddType $t (WithCurrentPrec p a) = WithCurrentPrec p (AddType $t a)
       add = liftT1 add
 
     instance
       (CanAddAsymmetric (CN $t) a)
       => 
-      CanAddAsymmetric (CN $t) (WithCurrentPrec a p)
+      CanAddAsymmetric (CN $t) (WithCurrentPrec p a)
       where
-      type AddType (CN $t) (WithCurrentPrec a p) = WithCurrentPrec (AddType (CN $t) a) p
+      type AddType (CN $t) (WithCurrentPrec p a) = WithCurrentPrec p (AddType (CN $t) a)
       add = liftT1 add
 
     instance
       (CanSub a $t)
       => 
-      CanSub (WithCurrentPrec a p) $t
+      CanSub (WithCurrentPrec p a) $t
       where
-      type SubType (WithCurrentPrec a p) $t = WithCurrentPrec (SubType a $t) p
+      type SubType (WithCurrentPrec p a) $t = WithCurrentPrec p (SubType a $t)
       sub = lift1T sub
 
     instance
       (CanSub a (CN $t))
       => 
-      CanSub (WithCurrentPrec a p) (CN $t)
+      CanSub (WithCurrentPrec p a) (CN $t)
       where
-      type SubType (WithCurrentPrec a p) (CN $t) = WithCurrentPrec (SubType a (CN $t)) p
+      type SubType (WithCurrentPrec p a) (CN $t) = WithCurrentPrec p (SubType a (CN $t))
       sub = lift1T sub
 
     instance
       (CanSub $t a)
       => 
-      CanSub $t (WithCurrentPrec a p)
+      CanSub $t (WithCurrentPrec p a)
       where
-      type SubType $t (WithCurrentPrec a p) = WithCurrentPrec (SubType $t a) p
+      type SubType $t (WithCurrentPrec p a) = WithCurrentPrec p (SubType $t a)
       sub = liftT1 sub
 
     instance
       (CanSub (CN $t) a)
       => 
-      CanSub (CN $t) (WithCurrentPrec a p)
+      CanSub (CN $t) (WithCurrentPrec p a)
       where
-      type SubType (CN $t) (WithCurrentPrec a p) = WithCurrentPrec (SubType (CN $t) a) p
+      type SubType (CN $t) (WithCurrentPrec p a) = WithCurrentPrec p (SubType (CN $t) a)
       sub = liftT1 sub
 
     instance
       (CanMulAsymmetric a $t)
       => 
-      CanMulAsymmetric (WithCurrentPrec a p) $t
+      CanMulAsymmetric (WithCurrentPrec p a) $t
       where
-      type MulType (WithCurrentPrec a p) $t = WithCurrentPrec (MulType a $t) p
+      type MulType (WithCurrentPrec p a) $t = WithCurrentPrec p (MulType a $t)
       mul = lift1T mul
 
     instance
       (CanMulAsymmetric a (CN $t))
       => 
-      CanMulAsymmetric (WithCurrentPrec a p) (CN $t)
+      CanMulAsymmetric (WithCurrentPrec p a) (CN $t)
       where
-      type MulType (WithCurrentPrec a p) (CN $t) = WithCurrentPrec (MulType a (CN $t)) p
+      type MulType (WithCurrentPrec p a) (CN $t) = WithCurrentPrec p (MulType a (CN $t))
       mul = lift1T mul
 
     instance
       (CanMulAsymmetric $t a)
       => 
-      CanMulAsymmetric $t (WithCurrentPrec a p)
+      CanMulAsymmetric $t (WithCurrentPrec p a)
       where
-      type MulType $t (WithCurrentPrec a p) = WithCurrentPrec (MulType $t a) p
+      type MulType $t (WithCurrentPrec p a) = WithCurrentPrec p (MulType $t a)
       mul = liftT1 mul
 
     instance
       (CanMulAsymmetric (CN $t) a)
       => 
-      CanMulAsymmetric (CN $t) (WithCurrentPrec a p)
+      CanMulAsymmetric (CN $t) (WithCurrentPrec p a)
       where
-      type MulType (CN $t) (WithCurrentPrec a p) = WithCurrentPrec (MulType (CN $t) a) p
+      type MulType (CN $t) (WithCurrentPrec p a) = WithCurrentPrec p (MulType (CN $t) a)
       mul = liftT1 mul
 
     instance
       (CanDiv a $t)
       => 
-      CanDiv (WithCurrentPrec a p) $t
+      CanDiv (WithCurrentPrec p a) $t
       where
-      type DivType (WithCurrentPrec a p) $t = WithCurrentPrec (DivType a $t) p
+      type DivType (WithCurrentPrec p a) $t = WithCurrentPrec p (DivType a $t)
       divide = lift1T divide
 
     instance
       (CanDiv a (CN $t))
       => 
-      CanDiv (WithCurrentPrec a p) (CN $t)
+      CanDiv (WithCurrentPrec p a) (CN $t)
       where
-      type DivType (WithCurrentPrec a p) (CN $t) = WithCurrentPrec (DivType a (CN $t)) p
+      type DivType (WithCurrentPrec p a) (CN $t) = WithCurrentPrec p (DivType a (CN $t))
       divide = lift1T divide
 
     instance
       (CanDiv $t a)
       => 
-      CanDiv $t (WithCurrentPrec a p)
+      CanDiv $t (WithCurrentPrec p a)
       where
-      type DivType $t (WithCurrentPrec a p) = WithCurrentPrec (DivType $t a) p
+      type DivType $t (WithCurrentPrec p a) = WithCurrentPrec p (DivType $t a)
       divide = liftT1 divide
 
     instance
       (CanDiv (CN $t) a)
       => 
-      CanDiv (CN $t) (WithCurrentPrec a p)
+      CanDiv (CN $t) (WithCurrentPrec p a)
       where
-      type DivType (CN $t) (WithCurrentPrec a p) = WithCurrentPrec (DivType (CN $t) a) p
+      type DivType (CN $t) (WithCurrentPrec p a) = WithCurrentPrec p (DivType (CN $t) a)
       divide = liftT1 divide
 
   |]))