+++ aern2-mp/changelog.md
@@ -1,5 +1,10 @@
 # Change log for aern2-mp
 
+* v 0.2.7 2021-06-XXXXX
+  * make WithCurrentPrec p (CN MPBall) a Field
+  * simplify WithCurrentPrec, swap its parameters
+* v 0.2.6 2021-05-29
+  * adapt to new ppow operations
 * v 0.2.5 2021-05-27
   * add generic multivalued select + Kleenean instances
   * WithCurrentPrec: add MixedTypesNumPrelude class instances